+++ clx_budget_management/models/sale_subscription.py
@@ -22,22 +22,22 @@ class SaleSubscription(models.Model):
                 <br/> Upsell Date : %s <br/>
              Upsell Amount : %s
                 .</p>""") % (
-                           budget_line_id.sol_id.id,
-                           budget_line_id.sol_id.display_name,
-                           budget_line_id.sol_id.order_id.id,
-                           budget_line_id.sol_id.order_id.name,
-                           budget_line_id.subscription_id.id,
-                           budget_line_id.subscription_id.code,
-                           fields.Date.today(),
-                           user.id,
-                           user.name,
-                           budget_line_id.start_date if
-                           budget_line_id.status ==
-                           'upsell' else 'NO Upsell',
-                           budget_line_id.price if
-                           budget_line_id.status ==
-                           'upsell' else 0.0
-                       ))
+                                                           budget_line_id.sol_id.id,
+                                                           budget_line_id.sol_id.display_name,
+                                                           budget_line_id.sol_id.order_id.id,
+                                                           budget_line_id.sol_id.order_id.name,
+                                                           budget_line_id.subscription_id.id,
+                                                           budget_line_id.subscription_id.code,
+                                                           fields.Date.today(),
+                                                           user.id,
+                                                           user.name,
+                                                           budget_line_id.start_date if
+                                                           budget_line_id.status ==
+                                                           'upsell' else 'NO Upsell',
+                                                           budget_line_id.price if
+                                                           budget_line_id.status ==
+                                                           'upsell' else 0.0
+                                                       ))
 
     def prepared_vals(self, line, sale_budget_created, start_date=False):
         """
@@ -53,9 +53,7 @@ class SaleSubscription(models.Model):
                                                                                        days=-1)
 
         subscription_line_id = line.subscription_id.recurring_invoice_line_ids. \
-            filtered(lambda x: x.product_id.id == line.product_id.id
-                               and x.price_unit == line.price_unit
-                               and x.start_date == line.start_date)
+            filtered(lambda x: x.so_line_id.id == line.id)
         if start_date and end_date and start_date <= today <= end_date:
             state = "active"
             active = True
@@ -75,8 +73,14 @@ class SaleSubscription(models.Model):
             'state': state,
             'active': active,
             'wholesale_price': line.wholesale_price,
-            'subscription_line_id': subscription_line_id.id
+            'subscription_line_id': subscription_line_id.id,
+            'upsell_down_sell_price': 0.0,
+            'final_report_price': line.subscription_id.recurring_total
         }
+        if line.order_id.subscription_management != 'create':
+            vals.update({
+                'final_report_price': 0.0
+            })
         return vals
 
     def create_or_update_budget_line(self, line, sale_budget):
@@ -100,20 +104,24 @@ class SaleSubscription(models.Model):
             raise ValidationError(_(
                 "Please check Configuration of the Month When End date is not set."
             ))
+        today = fields.Date.today()
+        after_adding_month_date = today + relativedelta(months=month_selection - 1)
+        difference_today_end_date = relativedelta(after_adding_month_date, today)
         if line:
             available_budget_line = budget_line.filtered(
                 lambda x: x.sol_id.id == line.id)
             if not available_budget_line:
                 vals = self.prepared_vals(line, sale_budget, line.start_date)
+                subscription_line_id = self.env['sale.subscription.line'].browse(vals.get('subscription_line_id'))
                 line_start_date = vals.get('start_date')
                 budget_line_id = self.env['sale.budget.line'].create(vals)
                 self.create_chatter_log(budget_line_id, user)
                 if line.order_id.subscription_management == 'create':
                     if not line.end_date:
-                        month_start_date = datetime.date(datetime.datetime.today().year,
-                                                         month_selection, 1)
-                        r = relativedelta(month_start_date, line_start_date)
-                        for i in range(0, r.months):
+                        # month_start_date = datetime.date(datetime.datetime.today().year,
+                        #                                  month_selection, 1)
+                        # r = relativedelta(month_start_date, line_start_date)
+                        for i in range(1, difference_today_end_date.months):
                             temp = line_start_date + relativedelta(months=1)
                             vals = self.prepared_vals(line, sale_budget, temp)
                             budget_line_id = self.env['sale.budget.line'].create(vals)
@@ -127,6 +135,24 @@ class SaleSubscription(models.Model):
                             budget_line_id = self.env['sale.budget.line'].create(vals)
                             line_start_date = budget_line_id.start_date
                             self.create_chatter_log(budget_line_id, user)
+                elif line.order_id.subscription_management == 'upsell' and not subscription_line_id.end_date:
+                    old_budget_line = budget_line.filtered(
+                        lambda x: x.partner_id.id == line.order_id.partner_id.id and x.sol_id.id != line.id)
+                    old_budget_line = old_budget_line.filtered(lambda x: x.start_date >= line.start_date and x.product_id.id == line.product_id.id)
+                    [bgt_line.write({'upsell_down_sell_price': line.price_unit,
+                                     'final_report_price': line.subscription_id.recurring_invoice_line_ids.filtered(lambda x:x.line_type == 'base').price_subtotal + line.price_unit}) for bgt_line in
+                     old_budget_line]
+                elif line.order_id.subscription_management == 'upsell' and subscription_line_id.end_date:
+                    old_budget_line = self.env['sale.budget.line'].search([
+                        ('partner_id', '=', line.order_id.partner_id.id),
+                        ('product_id', '=', line.product_id.id),
+                        ('sol_id', '!=', line.id),
+                        ('start_date', '>=', line.start_date),
+                        ('end_date', '<=', line.end_date),
+                        '|', ('active', '=', True), ('active', '=', False)])
+                    [bgt_line.write({'upsell_down_sell_price': line.price_unit,
+                                     'final_report_price': line.subscription_id.recurring_invoice_line_ids.filtered(lambda x:x.line_type == 'base').price_subtotal + line.price_unit}) for bgt_line in
+                     old_budget_line]
 
     @api.model
     def _create_sale_budget(self, sale_order=False):