+++ clx_invoice_policy/models/sale_subscription.py
@@ -154,7 +154,7 @@ class SaleSubscriptionLine(models.Model):
             'analytic_account_id': line.order_id.analytic_account_id.id,
             'analytic_tag_ids': [(6, 0, line.analytic_tag_ids.ids)],
             'line_type': self.line_type,
-            'sale_line_ids' : line.ids
+            'sale_line_ids': line.ids
         }
 
         if line.display_type:
@@ -201,4 +201,25 @@ class SaleSubscriptionLine(models.Model):
                          ) else 1
                 ) * product_qty,
             })
+            if self._context.get('manual'):
+                advance_num_month = line.order_id.clx_invoice_policy_id.num_of_month + 1
+                lang = line.order_id.partner_invoice_id.lang
+                format_date = self.env['ir.qweb.field.date'].with_context(
+                    lang=lang).value_to_html
+                if self._context.get('sol'):
+                    period_msg = _("Invoicing period: %s - %s") % (
+                        format_date(fields.Date.to_string(line.start_date), {}),
+                        format_date(fields.Date.to_string(self._context.get(
+                            'end_date')), {}))
+                else:
+                    period_msg = _("Invoicing period: %s - %s") % (
+                        format_date(fields.Date.to_string(self._context.get(
+                            'start_date')), {}),
+                        format_date(fields.Date.to_string(self._context.get(
+                            'end_date')), {}))
+                res.update({
+                    'price_unit': line.price_unit * advance_num_month if (self._context.get(
+                        'end_date') - line.start_date).days not in (30, 31) else line.price_unit,
+                    'name': period_msg
+                })
         return res