+++ clx_invoice_reports/__manifest__.py
@@ -3,7 +3,7 @@
 # See LICENSE file for full copyright & licensing details.
 {
     'name': 'CLX Reports',
-    'version': '13.3.0.0.10',
+    'version': '13.3.0.0.13',
     'summary': 'CLX Reports',
     'sequence': 1,
     'description': """ CLX Reports """,
@@ -14,8 +14,8 @@
                 'clx_invoice_policy'
                 ],
     'data': [
-        # 'views/report_templates.xml',
-        # 'views/report_invoice_document.xml',
+        'views/report_templates.xml',
+        'views/report_invoice_document.xml',
         'views/res_company_views.xml'
 
     ],