+++ clx_task_management/__manifest__.py
@@ -3,7 +3,7 @@
 # See LICENSE file for full copyright & licensing details.
 {
     'name': 'CLX Task Management',
-    'version': '13.3.0.0.21',
+    'version': '13.3.0.0.28',
     'summary': 'CLX Task Management',
     'sequence': 1,
     'description': """ CLX Task Management """,
@@ -13,7 +13,8 @@
     'depends': ['sale',
                 'project',
                 'sale_management',
-                'portal'
+                'portal',
+                'sale_crm'
                 ],
     'data': [
         'security/ir.model.access.csv',
@@ -32,6 +33,7 @@
         'views/project_views.xml',
         'views/sale_order_views.xml',
         'views/product_category_views.xml',
+        'views/sale_subscription_line_views.xml',
         'wizard/task_cancel_warning_wizard_views.xml'
     ],
     'demo': [