+++ clx_task_management/models/__init__.py
@@ -10,3 +10,5 @@ from . import sale_order_line
 from . import sale_order
 from . import res_config_settings
 from . import product_category
+from . import crm_lead
+from . import request_form_attachments