+++ clx_task_management/models/project.py
@@ -30,6 +30,16 @@ class ProjectProject(models.Model):
 class ProjectTask(models.Model):
     _inherit = 'project.task'
 
+    def _compute_sub_task(self):
+        domain = [('parent_id', '=', self.repositary_task_id.id)]
+        records = self.env['sub.task'].search(domain)
+        for record in records:
+            task = self.search([('name', '=', record.sub_task_name), ('parent_id', '=', self.id),
+                                ('project_id', '=', self.project_id.id)])
+            if task:
+                record.stage_id = task[0].stage_id.id
+        self.clx_sub_task_ids = records.ids
+
     repositary_task_id = fields.Many2one('main.task', string='Repository Task')
     sub_repositary_task_ids = fields.Many2many('sub.task',
                                                string='Repository Sub Task')
@@ -40,6 +50,11 @@ class ProjectTask(models.Model):
     team_members_ids = fields.Many2many('res.users', string="Team Members")
     clx_sale_order_id = fields.Many2one('sale.order', string='Sale order')
     clx_sale_order_line_id = fields.Many2one('sale.order.line', string="Sale order Item")
+    clx_sub_task_ids = fields.Many2many('sub.task', 'sub_task_project_task_clx', 'sub_task_id', 'task_id',
+                                        string="Sub Task",
+                                        compute='_compute_sub_task'
+                                        )
+    requirements = fields.Text(string='Requirements')
 
     def prepared_sub_task_vals(self, sub_task, main_task):
         """